events: Move room::create::RoomType to ruma-common
