+++ crates/ruma-common/src/lib.rs
@@ -12,6 +12,7 @@ pub mod power_levels;
 pub mod presence;
 pub mod push;
 pub mod receipt;
+pub mod room;
 pub mod thirdparty;
 mod time;
 pub mod to_device;
